// src/app/dota-shuffle/model/dota-player.model.ts
export interface DotaPlayerModel {
  id: number;
  label?: string;
  name: string;
  mmr: number;
  team: number;
  medal: string;
  maxSoloMmr?: boolean;
}

export interface BasicPlayerModel {
  id: number;
  mmr: number;
}

export interface TotalCombination {
  total: number;
  diffMrr: number[];
  combination: BasicPlayerModel[][];
}

export interface BestShuffleResult {
  best1: BasicPlayerModel[];
  best2: BasicPlayerModel[];
  best3: BasicPlayerModel[];
}